reorganize page components so sidebar does disappear while loading metabot settings pages
