+++ frontend/src/metabase/metabot/components/MetabotAdmin/MetabotAdminLayout.tsx
@@ -1,0 +1,14 @@
+import ErrorBoundary from "metabase/ErrorBoundary";
+import { AdminSettingsLayout } from "metabase/common/components/AdminLayout/AdminSettingsLayout";
+
+import { MetabotNavPane } from "./MetabotNavPane";
+
+export const MetabotAdminLayout = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => (
+  <AdminSettingsLayout sidebar={<MetabotNavPane />}>
+    <ErrorBoundary>{children}</ErrorBoundary>
+  </AdminSettingsLayout>
+);
